TeachBot/backend/requirements.txt
flask
flask-cors
paddleocr
paddlepaddle
fitz
pymupdf
google-generativeai
python-dotenv
